Fixed pet status polling job, Removed PetLocation - using PetStatusActivity instead, Updated Readme.md to use tabs in PaperUI control, Updated API call for PetStatus (now using only one call for all pets), some other renaming changes
