+++ bundles/org.openhab.binding.surepetcare/src/main/java/org/openhab/binding/surepetcare/internal/SurePetcareConstants.java
@@ -48,7 +48,7 @@ public class SurePetcareConstants {
     public static final String USERNAME = "username";
     public static final String PASSWORD = "password";
     public static final String REFRESH_INTERVAL_TOPOLOGY = "refresh_interval_topology";
-    public static final String REFRESH_INTERVAL_LOCATION = "refresh_interval_location";
+    public static final String REFRESH_INTERVAL_STATUS = "refresh_interval_status";
 
     // Bridge Channel Names
     public static final String BRIDGE_CHANNEL_ONLINE = "online";
