+++ bundles/org.openhab.binding.surepetcare/src/main/java/org/openhab/binding/surepetcare/internal/data/SurePetcarePetStatus.java
@@ -12,12 +12,6 @@
  */
 package org.openhab.binding.surepetcare.internal.data;
 
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-
 import com.google.gson.annotations.SerializedName;
 
 /**
@@ -29,139 +23,24 @@ import com.google.gson.annotations.SerializedName;
  */
 public class SurePetcarePetStatus {
 
-    public class Activity {
-
-        @SerializedName("tag_id")
-        private Integer tagId;
-        @SerializedName("device_id")
-        private Integer deviceId;
-        @SerializedName("user_id")
-        private Integer userId;
-        @SerializedName("where")
-        private Integer where;
-        @SerializedName("since")
-        private Date since;
-
-        public Integer getTagId() {
-            return tagId;
-        }
-
-        public void setTagId(Integer tagId) {
-            this.tagId = tagId;
-        }
-
-        public Integer getUserId() {
-            return userId;
-        }
-
-        public void setUserId(Integer userId) {
-            this.userId = userId;
-        }
-
-        public Integer getDeviceId() {
-            return deviceId;
-        }
-
-        public void setDeviceId(Integer deviceId) {
-            this.deviceId = deviceId;
-        }
-
-        public Integer getWhere() {
-            return where;
-        }
-
-        public void setWhere(Integer where) {
-            this.where = where;
-        }
-
-        public Date getSince() {
-            return since;
-        }
-
-        public void setSince(Date since) {
-            this.since = since;
-        }
-
-        public ZonedDateTime getLocationChanged() {
-            if (since != null) {
-                return since.toInstant().atZone(ZoneId.systemDefault()).withNano(0);
-            } else {
-                return null;
-            }
-        }
-
-    }
-
-    public class Feeding {
-
-        @SerializedName("tag_id")
-        private Integer tagId;
-        @SerializedName("device_id")
-        private Integer deviceId;
-        @SerializedName("change")
-        private List<Float> feedChange = new ArrayList<Float>();
-        @SerializedName("at")
-        private Date feedChangeAt;
-
-        public Integer getTagId() {
-            return tagId;
-        }
-
-        public void setTagId(Integer tagId) {
-            this.tagId = tagId;
-        }
-
-        public Integer getDeviceId() {
-            return deviceId;
-        }
-
-        public void setDeviceId(Integer deviceId) {
-            this.deviceId = deviceId;
-        }
-
-        public List<Float> getFeedChange() {
-            return feedChange;
-        }
-
-        public void setFeedChange(List<Float> feedChange) {
-            this.feedChange = feedChange;
-        }
-
-        public Date getAt() {
-            return feedChangeAt;
-        }
-
-        public void setAt(Date feedChangeAt) {
-            this.feedChangeAt = feedChangeAt;
-        }
-
-        public ZonedDateTime getZonedFeedChangeAt() {
-            if (feedChangeAt != null) {
-                return feedChangeAt.toInstant().atZone(ZoneId.systemDefault());
-            } else {
-                return null;
-            }
-        }
-    }
-
     @SerializedName("activity")
-    private Activity activity;
+    private SurePetcarePetActivity activity;
     @SerializedName("feeding")
-    private Feeding feeding;
+    private SurePetcarePetFeeding feeding;
 
-    public Activity getActivity() {
+    public SurePetcarePetActivity getActivity() {
         return activity;
     }
 
-    public void setActivity(Activity activity) {
+    public void setActivity(SurePetcarePetActivity activity) {
         this.activity = activity;
     }
 
-    public Feeding getFeeding() {
+    public SurePetcarePetFeeding getFeeding() {
         return feeding;
     }
 
-    public void setFeeding(Feeding feeding) {
+    public void setFeeding(SurePetcarePetFeeding feeding) {
         this.feeding = feeding;
     }
 }
