+++ bundles/org.openhab.binding.surepetcare/src/main/java/org/openhab/binding/surepetcare/internal/handler/SurePetcareBaseObjectHandler.java
@@ -49,7 +49,7 @@ public abstract class SurePetcareBaseObjectHandler extends BaseThingHandler {
 
     @Override
     public void handleCommand(ChannelUID channelUID, Command command) {
-        logger.debug("DeviceHandler handleCommand called with command: {}", command.toString());
+        logger.debug("BaseObjectHandler handleCommand called with command: {}", command.toString());
         if (command instanceof RefreshType) {
             updateThing();
         }
