+++ bundles/org.openhab.binding.surepetcare/src/main/resources/ESH-INF/i18n/surepetcare.properties
@@ -12,8 +12,8 @@ thing-type.config.surepetcare.bridge.username.description = The username to acce
 thing-type.config.surepetcare.bridge.password.label = Sure Petcare Password
 thing-type.config.surepetcare.bridge.password.description = The password to access the Sure Petcare API.
 
-thing-type.config.surepetcare.bridge.refresh_interval_location.label = Refresh Interval Location
-thing-type.config.surepetcare.bridge.refresh_interval_location.description = Query interval (in secs) for pet locations. (min. 300 / default 36000).
+thing-type.config.surepetcare.bridge.refresh_interval_status.label = Refresh Interval Pet Status
+thing-type.config.surepetcare.bridge.refresh_interval_status.description = Query interval (in secs) for pet status. (min. 300 / default 36000).
 
 thing-type.config.surepetcare.bridge.refresh_interval_topology.label = Refresh Interval Topology
 thing-type.config.surepetcare.bridge.refresh_interval_topology.description = Query interval (in secs) for device topology. (min 300 / default 300).
