+++ bundles/org.openhab.binding.surepetcare/src/main/resources/ESH-INF/i18n/surepetcare_de.properties
@@ -12,8 +12,8 @@ thing-type.config.surepetcare.bridge.username.description = Sure Petcare Benutze
 thing-type.config.surepetcare.bridge.password.label = Sure Petcare Passwort
 thing-type.config.surepetcare.bridge.password.description = Sure Petcare Passwort.
 
-thing-type.config.surepetcare.bridge.refresh_interval_location.label = Abfrageintervall Standort
-thing-type.config.surepetcare.bridge.refresh_interval_location.description = Intervall zur Abfrage des Haustier Standorts (in Sekunden) (mind. 300 / standard 36000).
+thing-type.config.surepetcare.bridge.refresh_interval_status.label = Abfrageintervall Haustier Status
+thing-type.config.surepetcare.bridge.refresh_interval_status.description = Intervall zur Abfrage des Haustier Status (in Sekunden) (mind. 300 / standard 36000).
 
 thing-type.config.surepetcare.bridge.refresh_interval_topology.label = Abfrageintervall Topology
 thing-type.config.surepetcare.bridge.refresh_interval_topology.description = Intervall zur Abfrage der Geräte und Haustier Daten (in Sekunden) (mind. 300 / standard 300).
