+++ bundles/org.openhab.binding.surepetcare/src/main/resources/ESH-INF/thing/bridge.xml
@@ -28,9 +28,9 @@
 				<description>Query interval (in secs) for device topology</description>
 				<default>36000</default>
 			</parameter>
-			<parameter name="refresh_interval_location" type="integer" required="false" min="30">
-				<label>Refresh Interval Location</label>
-				<description>Query interval (in secs) for pet locations</description>
+			<parameter name="refresh_interval_status" type="integer" required="false" min="60">
+				<label>Refresh Interval Pet Status</label>
+				<description>Query interval (in secs) for pet status</description>
 				<default>300</default>
 			</parameter>
